+++ lib/src/interface/IErrorModel.dart
@@ -4,4 +4,5 @@ abstract class IErrorModel<T> {
   int? statusCode;
   String? description;
   INetworkModel? model;
+  dynamic response;
 }